// src/app/models/asistencia.ts
export interface asistencia{
    idasistencia?: number;
    iduser?: number;
    date_enter?: string;
    date_start? :string;
    date_end? :string;
    aula? : number;
    Estado? : string;
    comments?: string;
    created_at?: string;
    updated_at?: string;
}
/*
Este código define una interfaz llamada "asistencia" que tiene varios campos opcionales y obligatorios, como "idasistencia",
"iduser", "date_enter", "date_start", "date_end", "aula", "Estado", "comments", "created_at" y "updated_at". Esta interfaz
probablemente se usa para representar los datos de asistencia de los usuarios.*/